Refactored BaptismData and Users classes and their respective layers

+++ src/main/java/com/br/saraweb20/service/UserService.java
@@ -1,13 +1,23 @@
 package com.br.saraweb20.service;
 
-import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
+import org.springframework.transaction.annotation.Transactional;
 
+import com.br.saraweb20.dto.RoleDTO;
+import com.br.saraweb20.dto.UserDTO;
+import com.br.saraweb20.dto.UserInsertDTO;
+import com.br.saraweb20.entities.Role;
 import com.br.saraweb20.entities.User;
+import com.br.saraweb20.repositories.RoleRepository;
 import com.br.saraweb20.repositories.UserRepository;
+import com.br.saraweb20.service.exceptions.DatabaseException;
 import com.br.saraweb20.service.exceptions.ResourceNotFoundException;
 
 import jakarta.persistence.EntityNotFoundException;
@@ -18,40 +28,70 @@ public class UserService {
 	@Autowired
 	private UserRepository repository;
 	
-	public List<User> findAll() {
-		return repository.findAll();
-	}
+	@Autowired
+	private RoleRepository roleRepository;
 	
-	public User findById(Long id) {
+	@Transactional(readOnly = true)
+	public Page<UserDTO> findAllPaged(Pageable pageable) {
+		Page<User> list = repository.findAll(pageable);
+		return list.map(x -> new UserDTO(x));
+	}
+
+	@Transactional(readOnly = true)
+	public UserDTO findById(Long id) {
 		Optional<User> obj = repository.findById(id);
-		return obj.orElseThrow(() -> new ResourceNotFoundException(id));
+		User entity = obj.orElseThrow(() -> new ResourceNotFoundException(id));
+		return new UserDTO(entity);
 	}
-	
-	public List<User> findByName(String name) {
-		return repository.findByName(name);
+
+	@Transactional
+	public UserDTO insert(UserInsertDTO dto) {
+		User entity = new User();
+		copyDtoToEntity(dto, entity);
+		entity.setPassword(dto.getPassword());
+		entity = repository.save(entity);
+		return new UserDTO(entity);
 	}
-	
-	public User insert(User data) {
-		return repository.save(data);
-	}
-	
-	public void delete(Long id) {
-		findById(id);
-		repository.deleteById(id);
-	}
-	
-	public User update(Long id, User obj) {
+
+	@Transactional
+	public UserDTO update(Long id, UserDTO dto) {
+		if(!repository.existsById(id)) {
+			throw new ResourceNotFoundException(id);
+		}
 		try {
 			User entity = repository.getReferenceById(id);
-			updateData(entity, obj);
-			return repository.save(entity);
-		} catch (EntityNotFoundException e) {
+			copyDtoToEntity(dto, entity);
+			entity = repository.save(entity);
+			return new UserDTO(entity);
+		} 
+		catch (EntityNotFoundException e) {
 			throw new ResourceNotFoundException(id);
+		}		
+	}
+
+	@Transactional(propagation = Propagation.SUPPORTS)
+	public void delete(Long id) {
+		if(!repository.existsById(id)) {
+			throw new ResourceNotFoundException(id);
+		}
+		try {
+			repository.deleteById(id);
+		}
+		catch (DataIntegrityViolationException exception) {
+			throw new DatabaseException(exception.getMessage());
 		}
 	}
 	
-	private void updateData(User entity, User obj) {
-		entity.setName(obj.getName());
-		entity.setCpf(obj.getCpf());
+	private void copyDtoToEntity(UserDTO dto, User entity) {
+
+		entity.setName(dto.getName());
+		entity.setCpf(dto.getCpf());
+		entity.setLogin(dto.getLogin());
+		
+		entity.getRoles().clear();
+		for (RoleDTO roleDto : dto.getRoles()) {
+			Role role = roleRepository.getReferenceById(roleDto.getId());
+			entity.getRoles().add(role);
+		}
 	}
 }
